pyproject.toml
[project]
name = "camera-organizer"
version = "0.1.0"
description = "GUI application for organizing camera files by date with preview and batch operations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
]
